pyproject.toml
[project]
name = "flash-eng-backend-server"
version = "0.1.0"
description = "AI-powered resume builder using GitHub + LinkedIn data"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "supabase>=2.0.0",
    "python-multipart>=0.0.18",
    "PyMuPDF>=1.25.0",
    "httpx>=0.28.0",
    "playwright>=1.49.0",
    "jinja2>=3.1.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "PyJWT[crypto]>=2.0.0",
    "abacatepay>=1.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
